Put listener attributes and methods into BaseOperation and tests of these into new test_base_operation.py
